test: strawman for our scalability testing rig
Contain a scalability testing rig for flux-sched.

This is the first commit to get some early feedback as
I start this as a means to execute our CTS-1 flux-sched
testing plan described in distribution issue #14.

As such, it is currently limited to support the Phase I
of our CTS-1 testing plan.

README.md describes more details.