sharness: add project sharness scripts
scal-setup.sh sets up an environment to run one instance in the experiment.
It depends on a set of environment variables that are passed in by the
upper level script. But if they are not passed, it tries to set
some default values.

scal-sharness.sh has a set of common shell routines that will be used
used by sharness tests.